pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.4",
    "aioredis>=2.0.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "google-cloud-storage>=3.1.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.5",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.34.2",
]
